Jdp200303 35

+++ src/main/java/com/kodilla/ecommercee/service/OrderService.java
@@ -1,16 +1,20 @@
 package com.kodilla.ecommercee.service;
 
 import com.kodilla.ecommercee.entity.UserOrder;
+import com.kodilla.ecommercee.exception.OrderNotFoundException;
+import com.kodilla.ecommercee.exception.OrderAlreadyResolved;
 import com.kodilla.ecommercee.repository.OrderRepository;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
+
 
 
 @Service
 public class OrderService {
 
+    @Autowired
     private OrderRepository orderRepository;
 
     public List<UserOrder> getOrders() {
@@ -21,11 +25,18 @@ public class OrderService {
         return orderRepository.save(order);
     }
 
-    public Optional<UserOrder> getOrder(Long orderId) {
-        return orderRepository.findById(orderId);
+    public UserOrder getOrder(Long orderId) throws OrderNotFoundException {
+        return orderRepository.findById(orderId).orElseThrow(OrderNotFoundException::new);
     }
 
-    public void deleteOrder(Long orderId) {
-        orderRepository.deleteById(orderId);
+    public void deleteOrder(Long orderId) throws OrderAlreadyResolved {
+        if(orderRepository.existsById(orderId)) {
+            if (!orderRepository.getOne(orderId).isResolved()) {
+                orderRepository.deleteById(orderId);
+            } else {
+                throw new OrderAlreadyResolved();
+            }
+        }
+
     }
 }
